# nestjs-project/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "description": "Creates a new user account and sends an email confirmation link.",
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "email": {
                      "type": "string",
                      "format": "email"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Email already registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "summary": "Register a new user",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/confirm-email": {
      "get": {
        "description": "Validates the confirmation token and marks the user's email as confirmed.",
        "operationId": "AuthController_confirmEmail",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Email confirmed successfully"
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired confirmation token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "summary": "Confirm email address",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/resend-confirmation": {
      "post": {
        "description": "Resends the email confirmation link. No-op if the email is already confirmed or does not exist.",
        "operationId": "AuthController_resendConfirmation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendConfirmationDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Confirmation email sent (or silently skipped if already confirmed)"
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "summary": "Resend confirmation email",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "description": "Authenticates a confirmed user and returns an access + refresh token pair.",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "access_token": {
                      "type": "string"
                    },
                    "refresh_token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Invalid email or password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Email not confirmed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "summary": "Log in",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "description": "Issues a new access + refresh token pair using a valid refresh token.",
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tokens refreshed",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "access_token": {
                      "type": "string"
                    },
                    "refresh_token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Invalid, expired, or reused refresh token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "summary": "Refresh tokens",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/forgot-password": {
      "post": {
        "description": "Sends a password reset email. No-op if the email is not registered.",
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Reset email sent (or silently skipped if email not found)"
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "summary": "Request password reset",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/reset-password": {
      "post": {
        "description": "Resets the user password using a valid reset token.",
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Password reset successfully"
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired reset token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "summary": "Reset password",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "description": "Revokes all active refresh tokens for the authenticated user.",
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Logged out successfully"
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Log out",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "description": "Returns the authenticated user's profile from the JWT payload.",
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current user payload",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "sub": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "email": {
                      "type": "string",
                      "format": "email"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get current user",
        "tags": [
          "auth"
        ]
      }
    },
    "/videos/uploads": {
      "post": {
        "description": "Pre-registers the video as a draft and returns presigned URLs for a multipart upload made directly to object storage. Video bytes never pass through this API.",
        "operationId": "VideosController_initiateUpload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitiateUploadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InitiateUploadResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "CHANNEL_NOT_FOUND"
          },
          "413": {
            "description": "VIDEO_TOO_LARGE"
          },
          "415": {
            "description": "UNSUPPORTED_VIDEO_FORMAT"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Start a video upload",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{id}/uploads/complete": {
      "post": {
        "description": "Finalises the multipart upload and queues the video for processing.",
        "operationId": "VideosController_completeUpload",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Internal video id (uuid)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteUploadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompleteUploadResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "VIDEO_NOT_FOUND"
          },
          "409": {
            "description": "INVALID_VIDEO_STATE"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Finish a video upload",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{id}/uploads": {
      "delete": {
        "description": "Aborts the multipart upload and discards the draft video.",
        "operationId": "VideosController_abortUpload",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Internal video id (uuid)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Upload aborted"
          },
          "404": {
            "description": "VIDEO_NOT_FOUND"
          },
          "409": {
            "description": "INVALID_VIDEO_STATE"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Abandon a video upload",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{publicId}": {
      "get": {
        "description": "Anonymous for a ready video. A video still being processed is returned only to the channel that owns it, so the owner can poll progress.",
        "operationId": "VideosController_findOne",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "VIDEO_NOT_FOUND"
          }
        },
        "summary": "Get a video by its public identifier",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{publicId}/stream": {
      "get": {
        "description": "Redirects to a short-lived presigned URL. The client issues Range requests directly to object storage, which answers 206 Partial Content, so playback starts without downloading the whole file.",
        "operationId": "VideosController_stream",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to the media URL"
          },
          "404": {
            "description": "VIDEO_NOT_FOUND / VIDEO_NOT_READY"
          }
        },
        "summary": "Stream a video",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{publicId}/download": {
      "get": {
        "description": "Redirects to a presigned URL signed with an attachment Content-Disposition, so the browser saves the file.",
        "operationId": "VideosController_download",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to the download URL"
          },
          "404": {
            "description": "VIDEO_NOT_FOUND / VIDEO_NOT_READY"
          }
        },
        "summary": "Download a video",
        "tags": [
          "videos"
        ]
      }
    }
  },
  "info": {
    "title": "StreamTube API",
    "description": "API REST do StreamTube",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {}
      },
      "ResendConfirmationDto": {
        "type": "object",
        "properties": {}
      },
      "LoginDto": {
        "type": "object",
        "properties": {}
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {}
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {}
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {}
      },
      "InitiateUploadDto": {
        "type": "object",
        "properties": {
          "filename": {
            "type": "string",
            "example": "my-holiday.mp4",
            "maxLength": 255
          },
          "contentType": {
            "type": "string",
            "example": "video/mp4",
            "description": "Must be one of the configured allowed MIME types"
          },
          "sizeBytes": {
            "type": "number",
            "example": 1073741824,
            "description": "Declared size in bytes; must not exceed the configured maximum"
          },
          "title": {
            "type": "string",
            "example": "My holiday",
            "maxLength": 255
          }
        },
        "required": [
          "filename",
          "contentType",
          "sizeBytes"
        ]
      },
      "InitiateUploadPartDto": {
        "type": "object",
        "properties": {
          "partNumber": {
            "type": "number",
            "example": 1
          },
          "url": {
            "type": "string",
            "description": "Presigned PUT URL for this part"
          }
        },
        "required": [
          "partNumber",
          "url"
        ]
      },
      "InitiateUploadResponseDto": {
        "type": "object",
        "properties": {
          "videoId": {
            "type": "string"
          },
          "publicId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "processing",
              "ready",
              "failed"
            ]
          },
          "uploadId": {
            "type": "string"
          },
          "partSizeBytes": {
            "type": "number"
          },
          "parts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InitiateUploadPartDto"
            }
          },
          "expiresIn": {
            "type": "number",
            "description": "Seconds until the part URLs expire"
          }
        },
        "required": [
          "videoId",
          "publicId",
          "status",
          "uploadId",
          "partSizeBytes",
          "parts",
          "expiresIn"
        ]
      },
      "CompletedPartDto": {
        "type": "object",
        "properties": {
          "partNumber": {
            "type": "number",
            "example": 1,
            "minimum": 1
          },
          "etag": {
            "type": "string",
            "example": "\"9a0364b9e99bb480dd25e1f0284c8555\""
          }
        },
        "required": [
          "partNumber",
          "etag"
        ]
      },
      "CompleteUploadDto": {
        "type": "object",
        "properties": {
          "uploadId": {
            "type": "string",
            "example": "zX1a2b3c..."
          },
          "parts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompletedPartDto"
            }
          }
        },
        "required": [
          "uploadId",
          "parts"
        ]
      },
      "CompleteUploadResponseDto": {
        "type": "object",
        "properties": {
          "videoId": {
            "type": "string"
          },
          "publicId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "processing",
              "ready",
              "failed"
            ]
          }
        },
        "required": [
          "videoId",
          "publicId",
          "status"
        ]
      },
      "VideoChannelDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "nickname",
          "name"
        ]
      },
      "VideoResponseDto": {
        "type": "object",
        "properties": {
          "publicId": {
            "type": "string",
            "example": "A1b2C3d4E5f"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "processing",
              "ready",
              "failed"
            ]
          },
          "durationSeconds": {
            "type": "number",
            "nullable": true
          },
          "width": {
            "type": "number",
            "nullable": true
          },
          "height": {
            "type": "number",
            "nullable": true
          },
          "thumbnailUrl": {
            "type": "string",
            "nullable": true
          },
          "channel": {
            "$ref": "#/components/schemas/VideoChannelDto"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "publicId",
          "title",
          "status",
          "channel",
          "createdAt"
        ]
      },
      "ApiErrorEnvelope": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 401
          },
          "error": {
            "type": "string",
            "example": "INVALID_CREDENTIALS"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "Invalid email or password"
          },
          "code": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "statusCode",
          "error",
          "message"
        ]
      }
    }
  }
}